Create Analyzing the Online Retail Dataset
In this project, we will be analyzing the "Online Retail" dataset, which contains information about transactions made by customers of an online retail company. We will be looking for data anomalies, such as duplicate transactions and invalid data, and we will be using the data to gain insights into customer behavior and product performance.

--Prerequisites--
Before running the code in this project, you will need to install the following libraries:
pandas
seaborn (optional, for data visualization)
---
You can install these libraries using pip:

pip install pandas seaborn
---
To run the code in this project, simply run the following command:

python analyze_online_retail.py

This will execute the code in the analyze_online_retail.py file, which will perform the data anomaly and business insight analysis on the "Online Retail" dataset.